Namespace the rad nav framework
